Removing template logic.
