+++ src/lib.rs
@@ -3,10 +3,8 @@ pub mod email;
 mod embedded;
 pub mod gui;
 pub mod oauth;
-pub mod templates;
 
 // Re-export commonly used items
 pub use args::Args;
 pub use email::send_email;
 pub use oauth::get_or_refresh_token;
-pub use templates::load_templates;
